requirements.txt
pandas
pyarrow
dbt-core
dbt-airflow
duckdb>=0.10.0
dbt-duckdb